Naming a constructor PS brakes the javascript no more.
Introduced test for psc-bundle.

 reference
