+++ tests/TestCompiler.hs
@@ -26,29 +26,21 @@ import Prelude.Compat
 
 import qualified Language.PureScript as P
 
-import Data.Char (isSpace)
+import Control.Arrow ((>>>))
 import Data.Function (on)
-import Data.List (sort, stripPrefix, intercalate, groupBy, sortBy, minimumBy)
+import Data.List (sort, stripPrefix, intercalate, minimumBy)
 import Data.Maybe (mapMaybe)
-import Data.Time.Clock (UTCTime())
 import qualified Data.Text as T
-import Data.Tuple (swap)
 
 import qualified Data.Map as M
 
 import Control.Monad
-import Control.Arrow ((***), (>>>))
-
-import Control.Monad.Reader
-import Control.Monad.Trans.Except
 
 import System.Exit
-import System.Process hiding (cwd)
+import System.Process
 import System.FilePath
-import System.Directory
 import System.IO
-import System.IO.UTF8
-import qualified System.FilePath.Glob as Glob
+import System.IO.UTF8 (readUTF8File)
 
 import TestUtils
 import Test.Tasty
@@ -59,29 +51,8 @@ main = testSpec "compiler" spec
 
 spec :: Spec
 spec = do
-
-  (supportModules, supportExterns, supportForeigns, passingTestCases, warningTestCases, failingTestCases) <- runIO $ do
-    cwd <- getCurrentDirectory
-    let passing = cwd </> "tests" </> "purs" </> "passing"
-    let warning = cwd </> "tests" </> "purs" </> "warning"
-    let failing = cwd </> "tests" </> "purs" </> "failing"
-    passingFiles <- getTestFiles passing <$> testGlob passing
-    warningFiles <- getTestFiles warning <$> testGlob warning
-    failingFiles <- getTestFiles failing <$> testGlob failing
-    ms <- getSupportModuleTuples
-    let modules = map snd ms
-    supportExterns <- runExceptT $ do
-      foreigns <- inferForeignModules ms
-      externs <- ExceptT . fmap fst . runTest $ P.make (makeActions modules foreigns) modules
-      return (externs, foreigns)
-    case supportExterns of
-      Left errs -> fail (P.prettyPrintMultipleErrors P.defaultPPEOptions errs)
-      Right (externs, foreigns) -> return (modules, externs, foreigns, passingFiles, warningFiles, failingFiles)
-
-  outputFile <- runIO $ do
-    tmp <- getTemporaryDirectory
-    createDirectoryIfMissing False (tmp </> logpath)
-    openFile (tmp </> logpath </> logfile) WriteMode
+  (supportModules, supportExterns, supportForeigns, [passingTestCases, warningTestCases, failingTestCases]) <- runIO $ setUpTests ["passing", "warning", "failing"]
+  outputFile <- runIO $ createOutputFile logfile 
 
   context "Passing examples" $
     forM_ passingTestCases $ \testPurs ->
@@ -104,35 +75,11 @@ spec = do
 
   where
 
-  -- A glob for all purs and js files within a test directory
-  testGlob :: FilePath -> IO [FilePath]
-  testGlob = Glob.globDir1 (Glob.compile "**/*.purs")
-
-  -- Groups the test files so that a top-level file can have dependencies in a
-  -- subdirectory of the same name. The inner tuple contains a list of the
-  -- .purs files and the .js files for the test case.
-  getTestFiles :: FilePath -> [FilePath] -> [[FilePath]]
-  getTestFiles baseDir
-    = map (filter ((== ".purs") . takeExtensions) . map (baseDir </>))
-    . groupBy ((==) `on` extractPrefix)
-    . sortBy (compare `on` extractPrefix)
-    . map (makeRelative baseDir)
-
   -- Takes the test entry point from a group of purs files - this is determined
   -- by the file with the shortest path name, as everything but the main file
   -- will be under a subdirectory.
   getTestMain :: [FilePath] -> FilePath
   getTestMain = minimumBy (compare `on` length)
-
-  -- Extracts the filename part of a .purs file, or if the file is in a
-  -- subdirectory, the first part of that directory path.
-  extractPrefix :: FilePath -> FilePath
-  extractPrefix fp =
-    let dir = takeDirectory fp
-        ext = reverse ".purs"
-    in if dir == "."
-       then maybe fp reverse $ stripPrefix ext $ reverse fp
-       else dir
 
   -- Scans a file for @shouldFailWith directives in the comments, used to
   -- determine expected failures
@@ -147,80 +94,8 @@ spec = do
   extractPragma :: String -> FilePath -> IO [String]
   extractPragma pragma = fmap go . readUTF8File
     where
-    go = lines >>> mapMaybe (stripPrefix ("-- @" ++ pragma ++ " ")) >>> map trim
+      go = lines >>> mapMaybe (stripPrefix ("-- @" ++ pragma ++ " ")) >>> map trim
 
-inferForeignModules
-  :: MonadIO m
-  => [(FilePath, P.Module)]
-  -> m (M.Map P.ModuleName FilePath)
-inferForeignModules = P.inferForeignModules . fromList
-  where
-    fromList :: [(FilePath, P.Module)] -> M.Map P.ModuleName (Either P.RebuildPolicy FilePath)
-    fromList = M.fromList . map ((P.getModuleName *** Right) . swap)
-
-trim :: String -> String
-trim = dropWhile isSpace >>> reverse >>> dropWhile isSpace >>> reverse
-
-modulesDir :: FilePath
-modulesDir = ".test_modules" </> "node_modules"
-
-makeActions :: [P.Module] -> M.Map P.ModuleName FilePath -> P.MakeActions P.Make
-makeActions modules foreigns = (P.buildMakeActions modulesDir (P.internalError "makeActions: input file map was read.") foreigns False)
-                               { P.getInputTimestamp = getInputTimestamp
-                               , P.getOutputTimestamp = getOutputTimestamp
-                               , P.progress = const (pure ())
-                               }
-  where
-  getInputTimestamp :: P.ModuleName -> P.Make (Either P.RebuildPolicy (Maybe UTCTime))
-  getInputTimestamp mn
-    | isSupportModule (P.runModuleName mn) = return (Left P.RebuildNever)
-    | otherwise = return (Left P.RebuildAlways)
-    where
-    isSupportModule = flip elem (map (P.runModuleName . P.getModuleName) modules)
-
-  getOutputTimestamp :: P.ModuleName -> P.Make (Maybe UTCTime)
-  getOutputTimestamp mn = do
-    let filePath = modulesDir </> T.unpack (P.runModuleName mn)
-    exists <- liftIO $ doesDirectoryExist filePath
-    return (if exists then Just (P.internalError "getOutputTimestamp: read timestamp") else Nothing)
-
-runTest :: P.Make a -> IO (Either P.MultipleErrors a, P.MultipleErrors)
-runTest = P.runMake P.defaultOptions
-
-compile
-  :: [P.Module]
-  -> [P.ExternsFile]
-  -> M.Map P.ModuleName FilePath
-  -> [FilePath]
-  -> ([P.Module] -> IO ())
-  -> IO (Either P.MultipleErrors [P.ExternsFile], P.MultipleErrors)
-compile supportModules supportExterns supportForeigns inputFiles check = runTest $ do
-  fs <- liftIO $ readInput inputFiles
-  ms <- P.parseModulesFromFiles id fs
-  foreigns <- inferForeignModules ms
-  liftIO (check (map snd ms))
-  let actions = makeActions supportModules (foreigns `M.union` supportForeigns)
-  case ms of
-    [singleModule] -> pure <$> P.rebuildModule actions supportExterns (snd singleModule)
-    _ -> P.make actions (supportModules ++ map snd ms)
-
-assert
-  :: [P.Module]
-  -> [P.ExternsFile]
-  -> M.Map P.ModuleName FilePath
-  -> [FilePath]
-  -> ([P.Module] -> IO ())
-  -> (Either P.MultipleErrors P.MultipleErrors -> IO (Maybe String))
-  -> Expectation
-assert supportModules supportExterns supportForeigns inputFiles check f = do
-  (e, w) <- compile supportModules supportExterns supportForeigns inputFiles check
-  maybeErr <- f (const w <$> e)
-  maybe (return ()) expectationFailure maybeErr
-
-checkMain :: [P.Module] -> IO ()
-checkMain ms =
-  unless (any ((== P.moduleNameFromString "Main") . P.getModuleName) ms)
-    (fail "Main module missing")
 
 checkShouldFailWith :: [String] -> P.MultipleErrors -> Maybe String
 checkShouldFailWith expected errs =
@@ -322,9 +197,6 @@ assertDoesNotCompile supportModules supportExterns supportForeigns inputFiles sh
 
   where
   noPreCheck = const (return ())
-
-logpath :: FilePath
-logpath = "purescript-output"
 
 logfile :: FilePath
 logfile = "psc-tests.out"
