+++ tests/purs/bundle/PSasConstructor.purs
@@ -1,0 +1,11 @@
+module Main where
+
+import Prelude
+import Effect (Effect)
+import Effect.Console (log)
+
+data P = PS 
+
+main :: Effect Unit
+main = do
+  log "Done"
